[memory] extract MemoryData class from memory_pool to another file
 - Extract MemoryData class from memory_pool.h and make memory_data.h file
 - Add doxygen
