+++ nntrainer/tensor/memory_data.h
@@ -1,0 +1,115 @@
+// SPDX-License-Identifier: Apache-2.0
+/**
+ * Copyright (C) 2022 Jiho Chu <jiho.chu@samsung.com>
+ *
+ * @file   memory_data.h
+ * @date   14 Oct 2022
+ * @see    https://github.com/nnstreamer/nntrainer
+ * @author Jiho Chu <jiho.chu@samsung.com>
+ * @bug    No known bugs except for NYI items
+ * @brief  MemoryData class
+ *
+ */
+
+#ifndef __MEMORY_DATA_H__
+#define __MEMORY_DATA_H__
+
+namespace nntrainer {
+
+using MemoryDataValidateCallback = std::function<void(unsigned int)>;
+
+/**
+ * @brief  MemoryData Class
+ */
+template <typename T = float> class MemoryData {
+public:
+  /**
+   * @brief  Constructor of Memory Data
+   * @param[in] addr Memory data
+   */
+  explicit MemoryData(T *addr) :
+    valid(true),
+    id(0),
+    address(addr),
+    validate_cb([](unsigned int) {}),
+    invalidate_cb([](unsigned int) {}) {}
+
+  /**
+   * @brief  Constructor of Memory Data
+   * @param[in] mem_id validate callback.
+   * @param[in] v_cb validate callback.
+   * @param[in] i_cb invalidate callback.
+   */
+  explicit MemoryData(unsigned int mem_id, MemoryDataValidateCallback v_cb,
+                      MemoryDataValidateCallback i_cb) :
+    valid(false),
+    id(mem_id),
+    address(nullptr),
+    validate_cb(v_cb),
+    invalidate_cb(i_cb) {}
+
+  /**
+   * @brief  Deleted constructor of Memory Data
+   */
+  explicit MemoryData() = delete;
+
+  /**
+   * @brief  Constructor of MemoryData
+   */
+  explicit MemoryData(MemoryDataValidateCallback v_cb,
+                      MemoryDataValidateCallback i_cb) = delete;
+  /**
+   * @brief  Constructor of MemoryData
+   */
+  explicit MemoryData(T *addr, MemoryDataValidateCallback v_cb,
+                      MemoryDataValidateCallback i_cb) = delete;
+
+  /**
+   * @brief  Destructor of Memory Data
+   */
+  virtual ~MemoryData() = default;
+
+  /**
+   * @brief  Set address
+   */
+  void setAddr(T *addr) { address = addr; }
+
+  /**
+   * @brief  Get address
+   */
+  T *getAddr() const { return address; }
+
+  /**
+   * @brief  Validate memory data
+   */
+  void validate() {
+    if (valid)
+      return;
+    validate_cb(id);
+  }
+
+  /**
+   * @brief  Invalidate memory data
+   */
+  void invalidate() {
+    if (!valid)
+      return;
+    invalidate_cb(id);
+  }
+
+  /**
+   * @brief  Set valid
+   */
+  void setValid(bool v) { valid = v; }
+
+private:
+  bool valid;
+  unsigned int id;
+  T *address;
+  MemoryDataValidateCallback validate_cb;
+  MemoryDataValidateCallback invalidate_cb;
+};
+
+} // namespace nntrainer
+
+#endif /* __MEMORY_DATA_H__ */
