+++ nntrainer/tensor/memory_pool.h
@@ -21,67 +21,12 @@
 
 #include <functional>
 #include <memory>
-#include <memory_planner.h>
 #include <vector>
 
+#include <memory_data.h>
+#include <memory_planner.h>
+
 namespace nntrainer {
-
-using MemoryDataValidateCallback = std::function<void(unsigned int)>;
-
-template <typename T = float> class MemoryData {
-public:
-  /**
-   * @brief  Constructor of Memory Data
-   * @param[in] addr Memory data
-   */
-  explicit MemoryData(T *addr) :
-    valid(true), id(0), address(addr), validate_cb([](unsigned int) {}), invalidate_cb([](unsigned int) {}) {}
-
-  /**
-   * @brief  Constructor of Memory Data
-   * @param[in] mem_id validate callback.
-   * @param[in] v_cb validate callback.
-   * @param[in] i_cb invalidate callback.
-   */
-  explicit MemoryData(unsigned int mem_id, MemoryDataValidateCallback v_cb, MemoryDataValidateCallback i_cb) :
-    valid(false), id(mem_id), address(nullptr), validate_cb(v_cb), invalidate_cb(i_cb) {}
-
-  /**
-   * @brief  Deleted constructor of Memory Data
-   */
-  explicit MemoryData() = delete;
-  explicit MemoryData(MemoryDataValidateCallback v_cb, MemoryDataValidateCallback i_cb) = delete;
-  explicit MemoryData(T *addr, MemoryDataValidateCallback v_cb, MemoryDataValidateCallback i_cb) = delete;
-
-  /**
-   * @brief  Destructor of Memory Data
-   */
-  virtual ~MemoryData() = default;
-
-  void setAddr(T *addr) { address = addr; }
-  T *getAddr() const { return address; }
-
-  void validate() {
-    if (valid)
-      return;
-    validate_cb(id);
-  }
-
-  void invalidate() {
-    if (!valid)
-      return;
-    invalidate_cb(id);
-  }
-
-  void setValid(bool v) { valid = v; }
-
-private:
-  bool valid;
-  unsigned int id;
-  T *address;
-  MemoryDataValidateCallback validate_cb;
-  MemoryDataValidateCallback invalidate_cb;
-};
 
 /**
  * @class   MemoryPool
@@ -112,9 +57,9 @@ public:
    * @note start_time is inclusive, but end_time is exclusive
    * @note The value of the return token starts from 1.
    */
-  virtual unsigned int requestMemory(size_t bytes, unsigned int start_time,
-                                     unsigned int end_time,
-                                     std::vector<unsigned int> exec_order = std::vector<unsigned int>());
+  virtual unsigned int requestMemory(
+    size_t bytes, unsigned int start_time, unsigned int end_time,
+    std::vector<unsigned int> exec_order = std::vector<unsigned int>());
 
   /**
    * @brief Plan the layout with memory planner
@@ -184,8 +129,19 @@ public:
   virtual bool isAllocated() const;
 
 protected:
+  /**
+   * @brief  Get memory offset
+   */
   std::vector<size_t> &getMemoryOffset() { return memory_offset; }
+
+  /**
+   * @brief  Get memory size
+   */
   std::vector<size_t> &getMemorySize() { return memory_size; }
+
+  /**
+   * @brief  Get memory execution order
+   */
   std::vector<std::vector<unsigned int>> &getMemoryExecOrder() {
     return memory_exec_order;
   }
